Eps6800 (#82)
* feat: eps6800, little endian

* feat: eps6800 improve

* feat: eps6800, treat status as model related

* feat: eps6800, treat status as model related

* feat: eps6800

* feat: eps6800

* feat: eps6800, fix RPT

* feat: eps6800, fix GLINT

* feat: eps6800, fix

* feat: eps6800, fix

* feat: eps6800, fix ax595 cursor

* feat: eps6800, segment lcd

* Revert "feat: eps6800, segment lcd"

This reverts commit 83e3c98b8b41346fc946c229a423c066717a17a9.

* feat: eps6009

* feat: eps6009

* Refactor unpacked nibbles check and error handling

* Revert "Refactor unpacked nibbles check and error handling"

This reverts commit 5af45bfd18c49ba977729d460b2678a01123154a.

* feat: eps6009 TBPTH/r0D fix

* feat: eps6009, remove data-lcd-* in reading board.svg

* feat: eps6009, svg optimize

* feat: eps6009, svg optimize

* feat: eps6009, fix keyboard

* feat: eps6009, fix keyboard

* feat: eps6009, fix eps6800

* feat: eps6009, review

* feat: eps6009, review

* feat: eps6009, ram

* feat: eps6009, review

* fix actions

---------

+++ CasioEmuMsvc/src/Gui/CallAnalysis.cpp
@@ -26,8 +26,10 @@ struct CallAnalysis : public UIWindow {
 	std::vector<FunctionCall> viewing_calls;
 	CallAnalysis() : UIWindow("Funcs") {
 		SetupHook(on_call_function, [this](casioemu::CPU& sender, const FunctionEventArgs& ea) {
-			const uint32_t xr0 = (sender.reg_r[3] << 24) | (sender.reg_r[2] << 16) |
-				(sender.reg_r[1] << 8) | sender.reg_r[0];
+			const uint32_t xr0 = (static_cast<uint32_t>(sender.reg_r[3]) << 24) |
+				(static_cast<uint32_t>(sender.reg_r[2]) << 16) |
+				(static_cast<uint32_t>(sender.reg_r[1]) << 8) |
+				static_cast<uint32_t>(sender.reg_r[0]);
 			OnCallFunction(ea.pc, ea.lr, xr0, sender.GetBacktrace());
 		});
 		SetupHook(on_eps_call_function, [this](const EpsFunctionEventArgs& ea) {
